refactor: rename shared to Shared
