+++ src/Pages/GlobalConfigurations/Authorization/APITokens/EditAPIToken.tsx
@@ -35,7 +35,7 @@ import {
     PermissionConfigurationForm,
     PermissionConfigurationFormProvider,
     usePermissionConfiguration,
-} from '../shared/components/PermissionConfigurationForm'
+} from '../Shared/components/PermissionConfigurationForm'
 import { createUserPermissionPayload, isDirectPermissionFormComplete } from '../utils'
 import { getDefaultUserStatusAndTimeout } from '../libUtils'
