+++ src/Pages/GlobalConfigurations/Authorization/PermissionGroups/List/PermissionGroupContainer.tsx
@@ -10,15 +10,15 @@ import React, { useRef } from 'react'
 import { importComponentFromFELibrary } from '../../../../../components/common'
 import { API_STATUS_CODES } from '../../../../../config'
 import { useAuthorizationContext } from '../../AuthorizationProvider'
-import FiltersEmptyState from '../../shared/components/FilterEmptyState/FilterEmptyState.component'
-import BulkSelectionActionWidget from '../../shared/components/BulkSelection/BulkSelectionActionWidget'
-import BulkSelectionModal from '../../shared/components/BulkSelection/BulkSelectionModal'
+import FiltersEmptyState from '../../Shared/components/FilterEmptyState/FilterEmptyState.component'
+import BulkSelectionActionWidget from '../../Shared/components/BulkSelection/BulkSelectionActionWidget'
+import BulkSelectionModal from '../../Shared/components/BulkSelection/BulkSelectionModal'
 import NoPermissionGroups from './NoPermissionGroups'
 import PermissionGroupListHeader from './PermissionGroupListHeader'
 import PermissionGroupTable from './PermissionGroupTable'
 import { PermissionGroupContainerProps } from './types'
-import { BulkSelectionModalTypes, useAuthorizationBulkSelection } from '../../shared/components/BulkSelection'
-import { BulkSelectionEntityTypes } from '../../shared/components/BulkSelection/constants'
+import { BulkSelectionModalTypes, useAuthorizationBulkSelection } from '../../Shared/components/BulkSelection'
+import { BulkSelectionEntityTypes } from '../../Shared/components/BulkSelection/constants'
 
 const PermissionGroupInfoBar = importComponentFromFELibrary('PermissionGroupInfoBar', noop, 'function')
