+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/AppPermissions/common.tsx
