+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/AppPermissions/index.ts
