+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/AppPermissions/utils.ts
