+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/ChartPermission/ChartPermission.component.tsx
