+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/ChartPermission/constants.ts
