+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/ChartPermission/index.ts
