+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/ChartPermission/types.ts
