+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/K8sObjectPermissions/K8sPermissions.component.tsx
