+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/K8sObjectPermissions/types.ts
