+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/PermissionConfigurationForm/types.ts
