+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/UserPermissionGroupsSelector/index.ts
