+++ src/Pages/GlobalConfigurations/Authorization/UserPermissions/List/UserPermissionContainer.tsx
@@ -8,15 +8,15 @@ import {
 } from '@devtron-labs/devtron-fe-common-lib'
 import React, { useRef } from 'react'
 import { API_STATUS_CODES } from '../../../../../config'
-import FiltersEmptyState from '../../shared/components/FilterEmptyState/FilterEmptyState.component'
+import FiltersEmptyState from '../../Shared/components/FilterEmptyState/FilterEmptyState.component'
 import NoUsers from './NoUsers'
 import { UserPermissionContainerProps } from './types'
 import UserPermissionListHeader from './UserPermissionListHeader'
-import BulkSelectionActionWidget from '../../shared/components/BulkSelection/BulkSelectionActionWidget'
-import BulkSelectionModal from '../../shared/components/BulkSelection/BulkSelectionModal'
+import BulkSelectionActionWidget from '../../Shared/components/BulkSelection/BulkSelectionActionWidget'
+import BulkSelectionModal from '../../Shared/components/BulkSelection/BulkSelectionModal'
 import UserPermissionTable from './UserPermissionTable'
-import { BulkSelectionModalTypes, useAuthorizationBulkSelection } from '../../shared/components/BulkSelection'
-import { BulkSelectionEntityTypes } from '../../shared/components/BulkSelection/constants'
+import { BulkSelectionModalTypes, useAuthorizationBulkSelection } from '../../Shared/components/BulkSelection'
+import { BulkSelectionEntityTypes } from '../../Shared/components/BulkSelection/constants'
 import { importComponentFromFELibrary } from '../../../../../components/common'
 
 const UserListFilterToolbar = importComponentFromFELibrary('UserListFilterToolbar', null, 'function')
