+++ src/Pages/GlobalConfigurations/Authorization/UserPermissions/List/UserPermissionTable.tsx
@@ -6,7 +6,7 @@ import {
 } from '@devtron-labs/devtron-fe-common-lib'
 import React from 'react'
 import { importComponentFromFELibrary } from '../../../../../components/common'
-import { useAuthorizationBulkSelection } from '../../shared/components/BulkSelection'
+import { useAuthorizationBulkSelection } from '../../Shared/components/BulkSelection'
 import { handleToggleCheckForBulkSelection } from '../../utils'
 import { userListLoading } from './constants'
 import { UserPermissionTableProps } from './types'
